API Documentation

DirectOrbits.KeplerianElements — Type

Orbit(
    a=1.0, # semi-major axis, AU
    i=π/2, # inclination, radians
    e=0.1, # eccentricity
    τ=π/2, # fraction of elements past periastron passage at MJD=0,
    μ=1.0, # graviational parameter, solar masses
    ω=π/2, # argument of periapsis
    Ω=π/2, # longitude of the ascending node
    plx=10.1, # paralax in milliarcseconds. Defines the distance to the object
)

Represents one object's Keplerian elementsal elements. Values can be specified by keyword argument for convinience, or kep2cart for efficiency.

See also KeplerianElementsDeg for a convinience constructor accepting units of degrees instead of radians.

DirectOrbits.astuple — Function

astuple(elements)

Return the parameters of a KeplerianElements value as a tuple.

DirectOrbits.KeplerianElementsDeg — Function

KeplerianElementsDeg(a, i, e, τ, μ, ω, Ω, plx)

A convinience function for constructing KeplerianElements where i, ω, and Ω are provided in units of degrees instead of radians.

Missing docstring.

Missing docstring for orbitsolve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for orbitsolve_ν. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OrbitSolution. Check Documenter's build log for details.

DirectOrbits.raoff — Function

raoff(elements, t)

Get the offset from the central body in Right Ascention in milliarcseconds at some time t in days.

DirectOrbits.decoff — Function

decoff(elements, t)

Get the offset from the central body in Declination in milliarcseconds at some time t in days.

DirectOrbits.period — Function

period(elem)

Period of an orbit in days.

DirectOrbits.distance — Function

distance(elem)

Distance to the system in parsecs.

Missing docstring.

Missing docstring for meanmotion. Check Documenter's build log for details.

DirectOrbits.periastron — Function

periastron(elements, tref=58849)

Compute the MJD of periastron passage most recently after the reference epoch tref. N.B. mjd of 58849 = 2020-01-01

DirectOrbits.radvel — Function

radvel(elements, t)

Get the radial velocity of the planet along the line of sight at the time t in days, in units of m/s.

radvel(elements, t, M_star, M_planet)

Get the radial velocity of the star along the line of sight at the time t in days, in units of m/s. The mass of the star and planet must have consistent units.

DirectOrbits.propmotionanom — Function

propmotionanom(elements, t, M_planet)

Calculate the instantenous proper motion anomaly on a star due to an orbiting companion.

Missing docstring.

Missing docstring for posangle. Check Documenter's build log for details.

DirectOrbits.projectedseparation — Function

projectedseparation(elements, t)

Projected separation in mas from the central body at time t (days).