API Documentation

DirectOrbits.KeplerianElements — Type

KeplerianElements(
    a=1.0, # semi-major axis [AU]
    i=π/2, # inclination [radians]
    e=0.1, # eccentricity
    τ=π/2, # epoch of periastron passage at MJD=0
    M=1.0, # mass of primary [M⊙]
    ω=π/2, # argument of periapsis [radians]
    Ω=π/2, # longitude of the ascending node [radians]
    plx=10.1, # parallax [milliarcseconds]; defines the distance to the object
)

Represents one object's Keplerian elements. Values can be specified by keyword argument or named tuple for convenience.

See also KeplerianElementsDeg for a convenience constructor accepting units of degrees instead of radians.

DirectOrbits.astuple — Function

astuple(elements)

Return the parameters of a KeplerianElements value as a tuple.

DirectOrbits.KeplerianElementsDeg — Function

KeplerianElementsDeg(a, i, e, τ, M, ω, Ω, plx)

A convenience function for constructing KeplerianElements where i, ω, and Ω are provided in units of degrees instead of radians.

DirectOrbits.orbitsolve — Function

orbitsolve(elements, t)

Given a set of orbital elements with a time t in days, get the position and velocity of the secondary body (e.g. planet around a star).

This will output an OrbitSolution struct with the following properties:

x: δ right ascension [milliarcseconds]

y: δ declination [milliarcseconds]

ẋ: right ascension proper motion anomaly [milliarcseconds/year]

ẏ: declination proper motion anomaly [milliarcseconds/year]

ż: radial velocity of the secondary [m/s]

You can access the properties by name .x. There are helper functions to calculate each of these properties individually, but if you need more than one it is most efficient to calculate them in one go.

radvel can optionally accept the mass of the primary to calculate the impact of the secondary body on radial velocity of the primary, instead of the radial velocity of the secondary body itself.

Note: these calculations use the small angle approximation, so are only accurate when the star is much further way from the observer than the secondary is from the primary.

See also: orbitsolve_ν, projectedseparation, raoff, decoff, radvel, propmotionanom.

DirectOrbits.orbitsolve_ν — Function

orbitsolve_ν(elem, ν)

Solve a keplerian orbit from a given true anomaly (in radians). See orbitsolve for the same function accepting a given time.

DirectOrbits.OrbitSolution — Type

OrbitSolution(
    x, # δ right ascension [milliarcseconds]
    y, # δ declination [milliarcseconds]
    ẋ, # right ascension proper motion anomaly [milliarcseconds/year]
    ẏ, # declination proper motion anomaly [milliarcseconds/year]
    ż, # radial velocity of the *secondary* [m/s]
    ẍ, # right ascension acceleration [milliarcseconds/year^2]
    ÿ, # declination acceleration [milliarcseconds/year^2]
)

Represents the secondary's position on the sky in terms of offset from the primary, its velocity and acceleration on the sky, and its radial velocity.

DirectOrbits.period — Function

period(elem)

Period of an orbit [days].

DirectOrbits.distance — Function

distance(elem)

Distance to the system [pc].

DirectOrbits.meanmotion — Function

meanmotion(elem)

Mean motion [rad/year].

DirectOrbits.periastron — Function

periastron(elements, tref=58849)

Compute the MJD of periastron passage most recently after the reference epoch tref. N.B. mjd of 58849 = 2020-01-01

DirectOrbits.raoff — Function

raoff(elem, t)

Get the offset [mas] from the primary body in Right Ascension at the time t [days].

raoff(o)

Get the offset [mas] from the primary body in Right Ascension from an instance of OrbitSolution.

DirectOrbits.decoff — Function

decoff(elem, t)

Get the offset [mas] from the primary body in Declination at the time t [days].

decoff(elem, t)

Get the offset [mas] from the primary body in Declination from an instance of OrbitSolution.

DirectOrbits.posangle — Function

posangle(elem, t)

Calculate the position angle [rad] of the secondary about its primary from our perspective at the time t [days].

posangle(o)

Calculate the position angle [rad] of the secondary about its primary from our perspective from an instance of OrbitSolution.

DirectOrbits.projectedseparation — Function

projectedseparation(elem, t)

Calculate the projected separation [mas] of the secondary from its primary at the time t [days].

projectedseparation(o)

Calculate the projected separation [mas] of the secondary from its primary from an instance of OrbitSolution.

DirectOrbits.propmotionanom — Function

propmotionanom(elem, t)

Get the instantaneous proper motion anomaly [mas/year] of the secondary at the time t [days].

propmotionanom(o)

Get the instantaneous proper motion anomaly [mas/year] of the secondary from an instance of OrbitSolution.

propmotionanom(elem, t, M_planet)

Get the instantaneous proper motion anomaly [mas/year] of the primary in at the time t [days]. The units of M_planet and elem.M must match.

DirectOrbits.radvel — Function

radvel(elem, t)

Get the radial velocity [m/s] of the secondary along the line of sight at the time t [days].

radvel(o)

Get the radial velocity [m/s] of the secondary along the line of sight from an instance of OrbitSolution.

radvel(elem, t, M_planet)

Get the radial velocity [m/s] of the primary along the line of sight at the time t [days]. The units of M_planet and elem.M must match.

DirectOrbits.acceleration — Function

acceleration(elem, t)

Get the instantaneous acceleration [mas/year^2] of the secondary at the time t [days].

acceleration(o)

Get the instantaneous acceleration [mas/year^2] of the secondary from an instance of OrbitSolution.

acceleration(elem, t, M_planet)

Get the instantaneous acceleration [mas/year^2] of the primary in at the time t [days]. The units of M_planet and elem.M must match.